* renaming the project
