+++ JasonState/Interfaces/IStateMachineExecutor.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using JasonState.Models;
+
+namespace JasonState.Interfaces
+{
+    public interface IStateMachineExecutor
+    {
+        void Execute(IEnumerable<BaseState> states);
+    }
+}
