+++ JasonState/JasonState.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>JasonState</AssemblyName>
+    <RootNamespace>JasonState</RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Flee" Version="1.2.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
+    <PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
+  </ItemGroup>
+</Project>
