+++ JasonState/Models/StateMachineModel.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace JasonState.Models
+{
+    public class StateMachineModel
+    {
+        public IEnumerable<StateModel> States { get; set; }
+    }
+}
