--- StateMachine.Core/Models/StateMachineContext.cs
@@ -1,9 +1,0 @@
-﻿using Flee.PublicTypes;
-
-namespace StateMachine.Core.Models
-{
-    internal static class StateMachineContext
-    {
-        public static ExpressionContext Context { get; set; } = new ExpressionContext();
-    }
-}
