--- StateMachine.Core/Models/StateMachineModel.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace StateMachine.Core.Models
-{
-    public class StateMachineModel
-    {
-        public IEnumerable<StateModel> States { get; set; }
-    }
-}
