--- StateMachine.sln
@@ -1,30 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2010
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StateMachine.TestClient", "StateMachine\StateMachine.TestClient.csproj", "{7DE6BCAA-B99D-4DDF-8524-42D4744F0845}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StateMachine.Core", "StateMachine.Core\StateMachine.Core.csproj", "{F67B6675-2AF9-4E89-AA9D-4B6227E360E7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7DE6BCAA-B99D-4DDF-8524-42D4744F0845}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7DE6BCAA-B99D-4DDF-8524-42D4744F0845}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7DE6BCAA-B99D-4DDF-8524-42D4744F0845}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7DE6BCAA-B99D-4DDF-8524-42D4744F0845}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F67B6675-2AF9-4E89-AA9D-4B6227E360E7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F67B6675-2AF9-4E89-AA9D-4B6227E360E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F67B6675-2AF9-4E89-AA9D-4B6227E360E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F67B6675-2AF9-4E89-AA9D-4B6227E360E7}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {97584178-FFFC-4475-9484-F30E4844065D}
-	EndGlobalSection
-EndGlobal
