--- StateMachine/Impls/States/ErrorState.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using StateMachine.Core.Models;
-
-namespace StateMachine.TestClient.Impls.States
-{
-    public class ErrorState : BaseState
-    {
-        public override void Execute()
-        {
-            Console.WriteLine("An error happened...");
-        }
-    }
-}
