--- StateMachine/Impls/States/ExecutePaymentState.cs
@@ -1,12 +1,0 @@
-﻿using StateMachine.Core.Models;
-
-namespace StateMachine.TestClient.Impls.States
-{
-    public class ExecutePaymentState : BaseState
-    {
-        public override void Execute()
-        {
-            throw new System.Exception();
-        }
-    }
-}
