--- StateMachine/Models/TestClientModel.cs
@@ -1,16 +1,0 @@
-﻿namespace StateMachine.TestClient.Models
-{
-    public static class TestClientModel
-    {
-        public static string FromEmail { get; set; }
-
-        public static string ToEmail { get; set; }
-
-        public static long CreditCardNumber { get; set; }
-
-        public static string CardHolderName { get; set; }
-
-        public static decimal Amount { get; set; }
-
-    }
-}
