+++ TestClient/Impls/States/ExecutePaymentState.cs
@@ -1,0 +1,12 @@
+﻿using JasonState.Models;
+
+namespace TestClient.Impls.States
+{
+    public class ExecutePaymentState : BaseState
+    {
+        public override void Execute()
+        {
+            throw new System.Exception();
+        }
+    }
+}
