+++ TestClient/Impls/States/FinalState.cs
@@ -1,0 +1,12 @@
+﻿using JasonState.Models;
+
+namespace TestClient.Impls.States
+{
+    public class FinalState : BaseState
+    {
+        public override void Execute()
+        {
+            System.Console.WriteLine("Final state executed.");
+        }
+    }
+}
