+++ TestClient/Impls/States/ValidatePaymentState.cs
@@ -1,0 +1,12 @@
+﻿using JasonState.Models;
+
+namespace TestClient.Impls.States
+{
+    public class ValidatePaymentState : BaseState
+    {
+        public override void Execute()
+        {
+            System.Console.WriteLine("Validate Payment state executed.");
+        }
+    }
+}
