+++ TestClient/Impls/Worker.cs
@@ -1,0 +1,43 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using JasonState.Interfaces;
+using JasonState.Models;
+using TestClient.Interfaces;
+
+namespace TestClient.Impls
+{
+    public class Worker : IWorker
+    {
+        private readonly IStateMachineBuilder _stateMachineBuilder;
+        private readonly IStateMachineExecutor _stateMachineExecutor;
+
+        public Worker(IStateMachineBuilder stateMachineBuilder, IStateMachineExecutor stateMachineExecutor)
+        {
+            _stateMachineBuilder = stateMachineBuilder;
+            _stateMachineExecutor = stateMachineExecutor;
+        }
+
+        public StateMachineModel LoadStateMachine()
+        {
+            var path = Path.Combine(Directory.GetCurrentDirectory(), @"Files/StateMachine.json");
+            
+            return _stateMachineBuilder.Load(path);
+        }
+
+        public IEnumerable<BaseState> BuildStateMachine(StateMachineModel stateMachine)
+        {
+            return _stateMachineBuilder.BuildMachine(stateMachine);
+        }
+
+        public void AddToContext(Type type)
+        {
+            _stateMachineBuilder.AddToContext(type);
+        }
+
+        public void Execute(IEnumerable<BaseState> machine)
+        {
+            _stateMachineExecutor.Execute(machine);
+        }
+    }
+}
