+++ TestClient/Models/TestClientModel.cs
@@ -1,0 +1,16 @@
+﻿namespace TestClient.Models
+{
+    public static class TestClientModel
+    {
+        public static string FromEmail { get; set; }
+
+        public static string ToEmail { get; set; }
+
+        public static long CreditCardNumber { get; set; }
+
+        public static string CardHolderName { get; set; }
+
+        public static decimal Amount { get; set; }
+
+    }
+}
